hover color change -- homepage
after an entire day, i've managed to get the background image of the site to change from neon yellow to gray on hover! although i intended to flesh out the html for each site page, i somehow ended up doing this. not really sure what direction i want to go in...complete the site page by page, or do it in layers (first all the html then all the css etc). this is progress though!